board 수정
