+++ src/main/java/com/team3webnovel/dao/BoardDao.java
@@ -1,15 +1,15 @@
 package com.team3webnovel.dao;
 
-import com.team3webnovel.vo.BoardVo;
 import java.util.List;
+import java.util.Map;
+
+import com.team3webnovel.vo.BoardVo;
 
 public interface BoardDao {
-    // 게시글 목록 가져오기
-    List<BoardVo> getAllPosts();
-    
-    // 게시글 ID로 단일 게시글 가져오기
-    BoardVo getPostById(int id);
-
-    // 게시글 추가
-    void addPost(BoardVo post);
+	public void insert (BoardVo board);
+	public BoardVo select(int boardId);
+	public int getTotal();
+	public List<BoardVo>boardPaging(Map <String, Integer> map);
+	public int delete(Map <String, Integer> map);
+	public void viewCount(int boardId);
 }
