+++ src/main/java/com/team3webnovel/services/BoardService.java
@@ -2,10 +2,10 @@ package com.team3webnovel.services;
 
 import com.team3webnovel.dto.BoardPageDto;
 import com.team3webnovel.dto.BoardViewDto;
-import com.team3webnovel.vo.GijeBoardVo;
+import com.team3webnovel.vo.BoardVo;
 
-public interface GijeBoardService {
-	public void write(GijeBoardVo vo);
+public interface BoardService {
+	public void write(BoardVo vo);
 	public BoardViewDto view(int boardId);
 	public BoardPageDto getPagingList(int page, int pageSize);
 	public String delete(int boardId, int userId);
