+++ src/main/resources/mappers/BoardMapper.xml
@@ -2,13 +2,13 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 	"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 	
-<mapper namespace="com.team3webnovel.mappers.GijeBoardMapper">
+<mapper namespace="com.team3webnovel.mappers.BoardMapper">
 
-	<insert id="write" parameterType="com.team3webnovel.vo.GijeBoardVo">
+	<insert id="write" parameterType="com.team3webnovel.vo.BoardVo">
 		insert into board (user_id, title, content) values (#{userId}, #{title}, #{content})
 	</insert>
 	
-	<select id="view" parameterType="int" resultType="com.team3webnovel.vo.GijeBoardVo">
+	<select id="view" parameterType="int" resultType="com.team3webnovel.vo.BoardVo">
 		select board_id as boardId, user_id as userId, title, content, created_at as createdAt, updated_at as updatedAt, view_count as viewCount
 		from board where board_id = #{boardId}
 	</select>
@@ -17,7 +17,7 @@
 	    SELECT COUNT(*) FROM board
 	</select>
 	
-	<select id="pagingBoardList" parameterType="map" resultType="com.team3webnovel.vo.GijeBoardVo">
+	<select id="pagingBoardList" parameterType="map" resultType="com.team3webnovel.vo.BoardVo">
 		<![CDATA[
 			SELECT * FROM (
 				SELECT a.*, ROWNUM rnum FROM (
