+++ src/main/webapp/static/js/board.js
@@ -1,4 +1,4 @@
 function goToPage(boardId, currentPage) {
-    const url = `/team3webnovel/gije/view/${boardId}?page=${currentPage}`;
+    const url = `/team3webnovel/board/view/${boardId}?page=${currentPage}`;
     window.location.href = url;
 }
